// src/app/museum/enums/department.ts
export enum Department {
    any_department = 0,
    american_decorative_arts = 1,
    ancient_near_eastern_art = 3,
    arms_and_armor = 4,
    arts_of_africa_oceania_and_the_americas = 5,
    asian_art = 6,
    cloisters = 7,
    costume_institute = 8,
    drawings_and_prints = 9,
    egyptian_art = 10,
    european_paintings = 11,
    european_sculpture_and_decorative_arts = 12,
    greek_and_roman_art = 13,
    islamic_art = 14,
    robert_lehman_collection = 15,
    libraries = 16,
    medieval_art = 17,
    musical_instruments = 18,
    photographs = 19,
    modern_art = 21
}