// src/app/museum/models/museum-object.ts
export class MuseumObject {
    constructor() {}

    objectID: number = -1;
    isHighlight?: boolean;
    accessionNumber?: string;
    accessionYear?: string;
    isPublicDomain?: boolean;
    primaryImage?: string;
    primaryImageSmall: string = "";
    additionalImages?: string[];
    constituents?: [{}];
    department?: string;
    objectName?: string;
    title: string = "";
    culture?: string;
    period?: string;
    dynasty?: string;
    reign?: string;
    portfolio?: string;
    artistRole?: string;
    artistPrefix?: string;
    artistDisplayName: string = "";
    artistDisplayBio?: string;
    artistSuffix?: string;
    artistAlphaSort?: string;
    artistNationality?: string;
    artistBeginDate?: string;
    artistEndDate?: string;
    artistGender?: string;
    artistWikidata_URL?: string;
    artistULAN_URL?: string;
    objectDate?: string;
    objectBeginDate?: number;
    objectEndDate?: number;
    medium?: string;
    dimensions?: string;
    dimensionsParsed?: [{}];
    measurements?: [{}];
    creditLine?: string;
    geographyType?: string;
    city?: string;
    state?: string;
    county?: string;
    country?: string;
    region?: string;
    subregion?: string;
    locale?: string;
    locus?: string;
    excavation?: string;
    river?: string;
    classification?: string;
    rightsAndReproduction?: string;
    linkResource?: string;
    metadataDate?: Date;
    repository?: string;
    objectURL?: string;
    tags?: [{}];
    objectWikidata_URL?: string;
    isTimelineWork?: boolean;
    GalleryNumber?: string;
}